pyproject.toml
[project]
name = "flask-orm-alchemy-utils-boiler-plate"
version = "1.0.0"
description = ""
authors = [
    {name = "Phyo Pyae Sone"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "gunicorn (>=24.1.1,<25.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pymysql (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
